# docs/assets/openapi.json
{
  "components": {
    "schemas": {
...
      "EventoNavegacao": {
        "properties": {
          "dados": {
            "additionalProperties": true,
            "title": "Dados",
            "type": "object"
          },
          "session_id": {
            "maxLength": 64,
            "minLength": 6,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Session Id",
            "type": "string"
          },
          "tipo": {
            "maxLength": 32,
            "title": "Tipo",
            "type": "string"
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          }
        },
        "required": [
          "session_id",
          "tipo"
        ],
        "title": "EventoNavegacao",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImovelPublico": {
        "description": "Resposta pública: mesmo imóvel + pontos de referência do bairro (dado que já existe em\n`sdr_shared.geo` para o agente entender \"perto da Faria Lima\" — o site só expõe o que já existia).",
        "properties": {
          "area_m2": {
            "title": "Area M2",
            "type": "number"
          },
          "bairro": {
            "title": "Bairro",
            "type": "string"
          },
          "cidade": {
            "title": "Cidade",
            "type": "string"
          },
          "condominio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condominio"
          },
          "descricao": {
            "title": "Descricao",
            "type": "string"
          },
          "destaque_investimento": {
            "default": false,
            "title": "Destaque Investimento",
            "type": "boolean"
          },
          "fotos": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Fotos",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "operacao": {
            "title": "Operacao",
            "type": "string"
          },
          "pontos_referencia": {
            "items": {
              "type": "string"
            },
            "title": "Pontos Referencia",
            "type": "array"
          },
          "preco": {
            "title": "Preco",
            "type": "number"
          },
          "quartos": {
            "title": "Quartos",
            "type": "integer"
          },
          "regiao": {
            "title": "Regiao",
            "type": "string"
          },
          "suites": {
            "default": 0,
            "title": "Suites",
            "type": "integer"
          },
          "tipo": {
            "title": "Tipo",
            "type": "string"
          },
          "vagas": {
            "default": 0,
            "title": "Vagas",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "tipo",
          "operacao",
          "cidade",
          "regiao",
          "bairro",
          "quartos",
          "area_m2",
          "preco",
          "descricao"
        ],
        "title": "ImovelPublico",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "api__routers__corretores__CorretorIn": {
        "properties": {
          "ativo": {
            "default": true,
            "title": "Ativo",
            "type": "boolean"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "foto": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foto"
          },
          "nome": {
            "minLength": 2,
            "title": "Nome",
            "type": "string"
          },
          "regioes": {
            "items": {
              "type": "string"
            },
            "title": "Regioes",
            "type": "array"
          },
          "telefone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefone"
          }
        },
        "required": [
          "nome"
        ],
        "title": "CorretorIn",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Token do painel": {
        "description": "Perfil local: o valor de `SDR_PAINEL_TOKEN` (em desenvolvimento, `dev-token`). Perfil AWS: o JWT emitido pelo Cognito.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Projeto Mora — Vértice Imóveis (POC acadêmica)"
    },
    "description": "\nAPI do **Mora**, o agente SDR imobiliário da Vértice Imóveis.\n\nA etiqueta de cada rota diz quem a usa e o que ela exige:\n\n| Etiqueta | Quem usa | Autenticação |\n|---|---|---|\n| `público` | site vitrine e chat | nenhuma |\n| `corretor` | painel do corretor | `Authorization: Bearer …` |\n| `operacao` | rotinas de atendimento no painel | `Authorization: Bearer …` |\n| `admin` | configuração, governança de IA e auditoria | `Authorization: Bearer …` |\n\n**Como autenticar no perfil local:** clique em **Authorize** e informe o valor de\n`SDR_PAINEL_TOKEN` (no ambiente de desenvolvimento, `dev-token`). No perfil AWS o token é o\nJWT do Cognito, validado pelo authorizer do API Gateway antes de chegar aqui.\n\nToda alteração feita por estas rotas é registrada em auditoria (ver `GET /auditoria`).\n",
    "license": {
      "name": "Uso acadêmico — licença a confirmar"
    },
    "title": "Mora — SDR Imobiliário API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auditoria": {
      "get": {
        "operationId": "listar_auditoria_get",
        "parameters": [
          {
            "in": "query",
            "name": "dias",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 365,
              "minimum": 1,
              "title": "Dias",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "ator",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ator"
            }
          },
          {
            "in": "query",
            "name": "acao",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Acao"
            }
          },
          {
            "in": "query",
            "name": "entidade",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entidade"
            }
          },
          {
            "in": "query",
            "name": "entidade_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entidade Id"
            }
          },
          {
            "in": "query",
            "name": "busca",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Busca"
            }
          },
          {
            "in": "query",
            "name": "so_sensiveis",
            "required": false,
            "schema": {
              "default": false,
              "title": "So Sensiveis",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 300,
              "maximum": 1000,
              "title": "Limite",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Listar",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/clientes": {
      "get": {
        "operationId": "listar_clientes_get",
        "parameters": [
          {
            "in": "query",
            "name": "busca",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Busca"
            }
          },
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "title": "Limite",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Listar",
        "tags": [
          "operacao"
        ]
      }
    },
...
    "/config": {
      "get": {
        "operationId": "obter_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Obter",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/corretores": {
      "get": {
        "operationId": "listar_corretores_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Listar",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "criar_corretores_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api__routers__corretores__CorretorIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Criar",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/dashboard/reativacao": {
      "get": {
        "description": "Fase 4 do ADR-0013: o que o aviso de imóvel novo produziu.\n\nJanela padrão de 30 dias porque reativação é lenta por natureza — o lead sumiu semanas atrás e o\nimóvel certo aparece quando aparece; sete dias mostrariam quase sempre zero.",
        "operationId": "reativacao_dashboard_reativacao_get",
        "parameters": [
          {
            "in": "query",
            "name": "dias",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 180,
              "minimum": 1,
              "title": "Dias",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Reativacao",
        "tags": [
          "corretor"
        ]
      }
    },
...
    "/eventos": {
      "post": {
        "operationId": "registrar_eventos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventoNavegacao"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Registrar",
        "tags": [
          "público"
        ]
      }
    },
...
    "/imoveis": {
      "get": {
        "operationId": "listar_imoveis_get",
        "parameters": [
          {
            "in": "query",
            "name": "operacao",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Operacao"
            }
          },
          {
            "in": "query",
            "name": "regiao",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Regiao"
            }
          },
          {
            "in": "query",
            "name": "preco_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Preco Max"
            }
          },
          {
            "in": "query",
            "name": "quartos",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quartos"
            }
          },
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 60,
              "maximum": 200,
              "title": "Limite",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ImovelPublico"
                  },
                  "title": "Response Listar Imoveis Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listar",
        "tags": [
          "público"
        ]
      }
    },
...
    "/leads": {
      "get": {
        "operationId": "listar_leads_get",
        "parameters": [
          {
            "in": "query",
            "name": "estagio",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Estagio"
            }
          },
          {
            "in": "query",
            "name": "temperatura",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Temperatura"
            }
          },
          {
            "in": "query",
            "name": "corretor_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Corretor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Listar",
        "tags": [
          "corretor"
        ]
      }
    },
...
    "/notificacoes": {
      "get": {
        "operationId": "listar_notificacoes_get",
        "parameters": [
          {
            "in": "query",
            "name": "apenas_nao_lidas",
            "required": false,
            "schema": {
              "default": false,
              "title": "Apenas Nao Lidas",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "title": "Limite",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Listar",
        "tags": [
          "operacao"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Perfil local (docker compose)",
      "url": "http://localhost:8000"
    },
    {
      "description": "Perfil AWS — a confirmar no deploy",
      "url": "https://api.exemplo.com.br"
    }
  ],
...
}